# tree-app/target/generated-sources/openapi/src/main/java/com/p3/app/api/NodesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.p3.app.api;

import com.p3.app.dto.CreateNodeRequest;
import com.p3.app.dto.NodeResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-18T15:24:57.413603700-06:00[America/Guatemala]")
@Validated
@Tag(name = "nodes", description = "the nodes API")
public interface NodesApi {

    /**
     * POST /nodes/{parentId}/children : Agregar un subordinado a un puesto existente
     *
     * @param parentId  (required)
     * @param createNodeRequest  (required)
     * @return Subordinado agregado. (status code 201)
     */
    @Operation(
        operationId = "addChild",
        summary = "Agregar un subordinado a un puesto existente",
        responses = {
            @ApiResponse(responseCode = "201", description = "Subordinado agregado.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NodeResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/nodes/{parentId}/children",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<NodeResponse> addChild(
        @Parameter(name = "parentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("parentId") Long parentId,
        @Parameter(name = "CreateNodeRequest", description = "", required = true) @Valid @RequestBody CreateNodeRequest createNodeRequest
    );


    /**
     * POST /nodes/root : Crear el puesto máximo (ej. Gerencia General)
     *
     * @param createNodeRequest  (required)
     * @return Raíz creada con éxito. (status code 201)
     */
    @Operation(
        operationId = "createRoot",
        summary = "Crear el puesto máximo (ej. Gerencia General)",
        responses = {
            @ApiResponse(responseCode = "201", description = "Raíz creada con éxito.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NodeResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/nodes/root",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<NodeResponse> createRoot(
        @Parameter(name = "CreateNodeRequest", description = "", required = true) @Valid @RequestBody CreateNodeRequest createNodeRequest
    );

}